LPD-25338 Move page tree test to site-navigation-admin-web
