--- modules/test/playwright/tests/layout-taglib/test.properties
@@ -1,5 +1,0 @@
-##
-## Testray
-##
-
-    testray.main.component.name=Page Administration
