+++ modules/test/playwright/tests/site-navigation-admin-web/pageSelector.spec.ts
@@ -9,9 +9,9 @@ import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
 import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
 import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../fixtures/loginTest';
+import {pageSelectorPagesTest} from '../../fixtures/pageSelectorPagesTest';
 import getRandomString from '../../utils/getRandomString';
-import {navigationMenusPagesTest} from '../site-navigation-admin-web/fixtures/navigationMenusPagesTest';
-import {pageSelectorPagesTest} from './fixtures/pageSelectorPagesTest';
+import {navigationMenusPagesTest} from './fixtures/navigationMenusPagesTest';
 
 export const test = mergeTests(
 	apiHelpersTest,
